Source React dependencies from NPM
As React is a fast moving library and new versions come out often it
seems like a good idea to be able to quickly and easily update the React
libraries that we depend on using a package manager.

Nominally we are using NPM here, however I'm actually using 'yarn' to do
the work. Yarn introduces a lock file that captures a snapshot of the
transient dependencies which is useful for making sure that whenever
someone checks out this repository they will always be able to get the
exact same set of dependencies that I used to develop this code with.

Using NPM/Yarn also obliges us to add a 'package.json'.

Finally, because we now have a lock file that lets us simply restore all
the packages just from the file we don't need to check in our
'node_modules' folder. Here I've added a '.gitignore' file that filters
these paths out.