Create the simplest possible Webpack configuration
The simplest possible Webpack configuration requires an entry point
which in our case is our very simple 'index.js', and it requires some
information on where to put the bundle it produces.

We are going to put our build output into a folder called 'BuildOutput'.
The contents of this folder will change every time that we run a build
so we'll also add an entry to our '.gitignore' so we never check in the
results.

Finally I'll update the reference in our web page.

At the moment we aren't doing anything interesting with our build but we
will soon.