Require CSS style using Webpack
It may seem quite an odd concept but Webpack lets us require assets
other than just other JavaScript modules. In this commit we add a
'require' for our CSS styles.

Again 'require'-ing items allows Webpack to process them and gives us
the opportunity for optimizations like pruning unused styles.

In this case our style is output now as an inline 'style' tag in the
web page so we can remove our link to the stylesheet for the moment.