Generate HTML via Webpack
It would be good if all our assets ended up in the 'BuildOutput' folder
so that we can start serving our website from there.

So in this commit I've add a Webpack plugin that generates the
'index.html' page from a template that we supply. It automatically adds
our bundled scripts to the page at the end of the body.