Split Webpack configuration file according to environment
In the future we want to be able to use some Webpack features only when
we are building our bundle for production and not for development.

For example optimizations like minification and tree shaking are
required for the production bundle but no for development.

Webpack expects that the configuration file exports either an object
with all required properties or a function that takes the environment
variable (pass to Webpack using the '--env' flag) and returns the
configuration object.

This gives us the opportunity to dynamically change our configuration
based on the environment value.

In this commit I've created a common, production and development
configuration. I'm also using 'webpack-merge' to merge the
configurations into the file object but you could also use
'Object.assign' or 'if' statements to achieve the same effect.

At the moment the production configuration does nothing but that's about
to change.