Visualize generated production bundle
It can be quite useful to see what's actually in your production bundle.
With the 'webpack-bundle-analyzer' you can see exactly what it's made up
of and how big those parts are.

Again I've added another config for visualizations and then created an
NPM script 'show' that lets me view the results in the browser.