Dynamically load a module using 'import()'
Webpack has a code splitting feature that allows us to dynamically load
bundles or chunks when needed. This can be very useful in a large
application where the code is naturally modular and we define a useful
point for importing the chunks or bundles, such as a router.

The 'import()' is supported by Babel through a plugin that parses it
only. Webpack then implements support for actually loading dynamically.